folder structure: refactor and split to plugin modules
